+++ distribution/release-notes.md
@@ -1,10 +1,18 @@
+## Elite Intel v-0.0365-beta
+
+### bug fixes / features:
+- Changed carrier estimated operation cost calculation
+- Added logging for Google TTS for debugging.
+
+
+
+
+
+
 ## Elite Intel v-0.0364-beta
 
 ### bug fixes / features:
 - Keybinding monitor fix. A) will not update the bindings without restart. B) it will not read the StartPreset, not just the bindings file that was updated last.
-
-
-
 
 
 ## Elite Intel v-0.0363-beta